+++ app/controllers/api/members_controller.rb
@@ -3,7 +3,7 @@
 # API Controller for resources of type User with role 'member'
 class API::MembersController < API::ApiController
   before_action :authenticate_user!, except: [:last_subscribed]
-  before_action :set_member, only: %i[update destroy merge complete_tour]
+  before_action :set_member, only: %i[update destroy merge complete_tour update_role]
   respond_to :json
 
   def index
@@ -200,6 +200,35 @@ class API::MembersController < API::ApiController
 
       render json: { tours: @member.profile.tours.split }
     end
+  end
+
+  def update_role
+    authorize @member
+
+    # we do not allow dismissing a user to a lower role
+    if params[:role] == 'member'
+      render 403 and return if @member.role == 'admin' || @member.role == 'manager'
+    elsif params[:role] == 'manager'
+      render 403 and return if @member.role == 'admin'
+    end
+
+    # do nothing if the role does not change
+    render json: @member and return if params[:role] == @member.role
+
+    ex_role = @member.role.to_sym
+    @member.remove_role ex_role
+    @member.add_role params[:role]
+
+    NotificationCenter.call type: 'notify_user_role_update',
+                            receiver: @member,
+                            attached_object: @member
+
+    NotificationCenter.call type: 'notify_admins_role_update',
+                            receiver: User.admins_and_managers,
+                            attached_object: @member,
+                            meta_data: { ex_role: ex_role }
+
+    render json: @member
   end
 
   private